Replacement de la data et mise en place du css + réorganisation des dossiers
