+++ frontend/src/App.scss
@@ -4,8 +4,8 @@
 @import "./pages/Signin/Signin.scss";
 @import "./pages/Intro_2/PageIntro2.scss";
 @import "./pages/user/Profil/Profil.scss";
-@import "./pages/PageEtape3/PageEtape3.scss";
-@import "./pages/PageEtape4/PageEtape4.scss";
+@import "./pages/user/PageEtape3/PageEtape3.scss";
+@import "./pages/user/PageEtape4/PageEtape4.scss";
 
 body {
   margin: 0;
