+++ frontend/src/pages/Intro_2/PageIntro2.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AbbePierre from "../../assets/Abbe-Pierre.png";
+import AbbePierre from "../../assets/abbe-pierre.png";
 
 function PageIntro2() {
   return (
